requirements.txt
robotframework==7.0.1
requests==2.31.0
psycopg2-binary==2.9.9
openpyxl==3.1.5
xmltodict==0.13.0
python-dotenv==1.0.0

